src/privacy_security.py
```python
import jwt
from cryptography.fernet import Fernet
from src.config import API_KEYS

# Encryption key for Fernet
ENCRYPTION_KEY = Fernet.generate_key()

def encrypt_data(data):
    """
    Function to encrypt user data
    """
    cipher_suite = Fernet(ENCRYPTION_KEY)
    encrypted_data = cipher_suite.encrypt(data.encode())
    return encrypted_data

def decrypt_data(encrypted_data):
    """
    Function to decrypt user data
    """
    cipher_suite = Fernet(ENCRYPTION_KEY)
    decrypted_data = cipher_suite.decrypt(encrypted_data).decode()
    return decrypted_data

def generate_jwt(user_id):
    """
    Function to generate JWT for user authentication
    """
    token = jwt.encode({"user_id": user_id}, API_KEYS['jwt_secret'], algorithm="HS256")
    return token

def verify_jwt(token):
    """
    Function to verify JWT for user authentication
    """
    try:
        decoded_token = jwt.decode(token, API_KEYS['jwt_secret'], algorithms=["HS256"])
        return decoded_token['user_id']
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None

def ensurePrivacy(data):
    """
    Function to ensure privacy of user data
    """
    encrypted_data = encrypt_data(data)
    return encrypted_data
```